style: fix typo

+++ module-infrastructure/src/main/kotlin/universe/sparkle/infrastructure/config/JwtConfig.kt
@@ -6,6 +6,6 @@ import org.springframework.context.annotation.Configuration
 @Configuration
 data class JwtConfig(
     @Value("\${jwt.secret}") val secret: String,
-    @Value("\${jwt.accessExpiryPeriodDay}") val accessExpiryPeriodDay: Int,
-    @Value("\${jwt.refreshExpiryPeriodDay}") val refreshExpiryPeriodDay: Int,
+    @Value("\${jwt.accessExpiryPeriodDay}") val accessExpiryPeriodDay: Long,
+    @Value("\${jwt.refreshExpiryPeriodDay}") val refreshExpiryPeriodDay: Long,
 )
